build: upgrade deps & minor format fixes
