+++ package.json
@@ -21,77 +21,87 @@
   },
   "license": "MIT",
   "dependencies": {
-    "@headlessui/react": "^2.1.9",
-    "@heroicons/react": "^2.1.5",
+    "@headlessui/react": "^2.2.0",
+    "@heroicons/react": "^2.2.0",
     "@svgr/webpack": "8.1.0",
     "@tailwindcss/aspect-ratio": "0.4.2",
     "@tailwindcss/container-queries": "^0.1.1",
     "@tailwindcss/forms": "0.5.9",
     "@tailwindcss/typography": "0.5.15",
     "autoprefixer": "10.4.20",
-    "axios": "^1.7.7",
+    "axios": "^1.7.8",
+    "axios-rate-limit": "^1.4.0",
     "bowser": "^2.11.0",
     "date-fns": "^4.1.0",
     "formik": "^2.4.6",
     "intl": "1.2.5",
     "moment": "^2.30.1",
     "next": "^14.2.14",
-    "nextjs-toploader": "^3.6.15",
-    "postcss": "8.4.47",
+    "nextjs-toploader": "^3.7.15",
+    "node-cache": "^5.1.2",
+    "plex-api": "^5.3.2",
+    "postcss": "8.4.49",
     "react": "^18.3.1",
     "react-animate-height": "^3.2.3",
-    "react-aria": "3.35.0",
-    "react-big-calendar": "^1.15.0",
-    "react-datepicker": "^7.4.0",
+    "react-aria": "3.36.0",
+    "react-big-calendar": "^1.16.3",
+    "react-datepicker": "^7.5.0",
     "react-dom": "18.3.1",
     "react-hot-toast": "^2.4.1",
-    "react-intl": "6.7.0",
+    "react-intl": "7.0.1",
     "react-popper-tooltip": "^4.4.2",
     "react-use-clipboard": "1.0.9",
-    "sass": "^1.79.4",
-    "tailwindcss": "3.4.13",
-    "theme-change": "^2.5.0"
+    "reflect-metadata": "^0.2.2",
+    "sass": "^1.81.0",
+    "swr": "^2.2.5",
+    "tailwindcss": "3.4.15",
+    "theme-change": "^2.5.0",
+    "typeorm": "^0.3.20",
+    "web-push": "^3.6.7",
+    "winston": "^3.17.0",
+    "winston-daily-rotate-file": "^5.0.0"
   },
   "devDependencies": {
-    "@commitlint/cli": "19.5.0",
-    "@commitlint/config-conventional": "19.5.0",
-    "@eslint/compat": "^1.2.0",
-    "@eslint/js": "^9.12.0",
+    "@commitlint/cli": "19.6.0",
+    "@commitlint/config-conventional": "19.6.0",
+    "@eslint/compat": "^1.2.3",
+    "@eslint/js": "^9.16.0",
     "@semantic-release/changelog": "6.0.3",
     "@semantic-release/commit-analyzer": "13.0.0",
     "@semantic-release/exec": "6.0.3",
     "@semantic-release/git": "10.0.1",
-    "@types/node": "^22.7.4",
-    "@types/react": "18.3.11",
-    "@types/react-big-calendar": "^1",
-    "@types/react-dom": "18.3.0",
+    "@types/node": "^22.10.1",
+    "@types/react": "18.3.12",
+    "@types/react-big-calendar": "^1.16.0",
+    "@types/react-dom": "18.3.1",
     "@types/react-transition-group": "4.4.11",
     "@types/semver": "7.5.8",
-    "@typescript-eslint/eslint-plugin": "8.8.0",
-    "@typescript-eslint/parser": "8.8.0",
+    "@types/web-push": "^3.6.4",
+    "@typescript-eslint/eslint-plugin": "8.16.0",
+    "@typescript-eslint/parser": "8.16.0",
     "commitizen": "4.3.1",
     "cy-mobile-commands": "^0.3.0",
-    "cypress": "^13.15.0",
+    "cypress": "^13.16.0",
     "cz-conventional-changelog": "3.3.0",
-    "daisyui": "^4.12.12",
-    "eslint": "^9.12.0",
-    "eslint-config-next": "^14.2.14",
+    "daisyui": "^4.12.14",
+    "eslint": "^9.16.0",
+    "eslint-config-next": "^15.0.3",
     "eslint-config-prettier": "9.1.0",
-    "eslint-plugin-formatjs": "5.0.0",
-    "eslint-plugin-jsx-a11y": "6.10.0",
+    "eslint-plugin-formatjs": "5.2.5",
+    "eslint-plugin-jsx-a11y": "6.10.2",
     "eslint-plugin-no-relative-import-paths": "1.5.5",
     "eslint-plugin-prettier": "5.2.1",
-    "eslint-plugin-react": "7.37.1",
-    "eslint-plugin-react-hooks": "4.6.2",
+    "eslint-plugin-react": "7.37.2",
+    "eslint-plugin-react-hooks": "5.0.0",
     "extract-react-intl-messages": "4.1.1",
-    "globals": "^15.10.0",
+    "globals": "^15.12.0",
     "lint-staged": "15.2.10",
-    "prettier": "3.3.3",
+    "prettier": "3.4.1",
     "prettier-plugin-organize-imports": "4.1.0",
-    "prettier-plugin-tailwindcss": "0.6.8",
-    "semantic-release": "24.1.2",
-    "stylelint": "^16.9.0",
-    "typescript": "^5.6.2"
+    "prettier-plugin-tailwindcss": "0.6.9",
+    "semantic-release": "24.2.0",
+    "stylelint": "^16.11.0",
+    "typescript": "^5.7.2"
   },
   "config": {
     "commitizen": {
